zagents/python_express_executor/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "python_express_executor"
version = "0.1.0"
dependencies = [
  "aiqtoolkit[langchain]",
]
requires-python = ">=3.11,<3.13"
description = "Custom AIQ Toolkit Workflow"
classifiers = ["Programming Language :: Python"]



[project.entry-points.'aiq.components']
python_express_executor = "python_express_executor.register"